. makeRootBinding
