+++ test/f/base.js
@@ -1,7 +1,9 @@
 const assert = require("assert");
 const base = require("../../src/f/base.js");
 
-describe("base");
+describe("base", function () {
+
+});
 
 describe("makeRootBinding", function() {
   // , including functions from base namespace class"
